Refactor size return type to be Int

+++ src/main/kotlin/com/pavleprica/kotlin/cache/Cache.kt
@@ -4,7 +4,7 @@ import java.util.*
 
 interface Cache<T, E> {
 
-    val size: Long
+    val size: Int
 
     operator fun set(key: T, value: E)
